Change amount type to long

+++ sentinel-generator/src/main/java/dev/delath/sentinel/domain/Transaction.java
@@ -1,14 +1,13 @@
 package dev.delath.sentinel.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
 
 public record Transaction(
         @JsonProperty("transaction_id") String transactionId,
         @JsonProperty("user_id") String userId,
         @JsonProperty("card_token") String cardToken,
         @JsonProperty("card_bin") String cardBin,
-        BigDecimal amount,
+        long amount,
         String currency,
         @JsonProperty("merchant_details") MerchantDetails merchantDetails,
         Location location,
